Rename temporal derivative to temporal solver
