+++ src/qm_sim/hamiltonian/temporal_solver/__init__.py
@@ -1,5 +1,5 @@
 from .base import get_temporal_solver
-from .base import BaseTemporalDerivative as TemporalDerivative
+from .base import BaseTemporalSolver as TemporalSolver
 
 # Tell python where to find subclasses
 from .crank_nicolson import CrankNicolson
